Cluster contextualization
User may optionally provide additional information in the 'ucernvm'
section of the context file to use a master/slave cluster
contextualization.

Master sends its IP address to the central service (cernvm-online.cern.ch),
from where it is fetched by the slave machines. The slave machines
are waiting for the master to publish the data (blocking wait in the
bootloader).